dhis2: fix unit tests
