+++ packages/dhis2/src/index.js
@@ -1,4 +1,3 @@
 import * as Adaptor from './Adaptor';
 export default Adaptor;
-
 export * from './Adaptor';
