+++ packages/dhis2/tsconfig.json
@@ -1,21 +1,13 @@
 {
   "include": ["src/**/*.ts"],
-  "exclude": ["node_modules"],
+  "exclude": ["node_modules", "*.d.ts"],
   "compilerOptions": {
+    "outDir": "dist-tmp",
     "allowJs": true,
     "esModuleInterop": true,
     "isolatedModules": true,
     "moduleResolution": "node",
-    "noImplicitAny": false,
-    "preserveConstEnums": true,
-    "removeComments": false,
-    "resolveJsonModule": true,
-    "sourceMap": false,
-    "strictNullChecks": false,
     "target": "ES2020",
     "module": "es2020"
-    // "declaration": true,
-    // "declarationDir": "types",
-    // "emitDeclarationOnly": true
   }
 }
